pyproject.toml
[project]
name = "youtube-orchestrator"
version = "0.1.0"
description = "Automated YouTube content pipeline"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.9.0",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "pydantic>=2.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "structlog>=24.0",
    "python-dotenv>=1.0",
    "click>=8.0",
    "google-api-python-client>=2.192.0",
    "anthropic>=0.84.0",
    "google-auth-oauthlib>=1.3.0",
    "elevenlabs>=2.39.0",
    "pillow>=12.1.1",
    "openai>=2.28.0",
    "google-auth>=2.49.1",
    "moviepy>=1.0.3",
    "opencv-python-headless>=4.13.0.92",
    "faster-whisper>=1.2.1",
    "praw>=7.8.0",
    "yt-dlp>=2024.0",
    "google-genai>=1.67.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "packages", "admin"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
